+++ lib/lib-storage/src/runtimeConfig.shared.ts
@@ -1,4 +1,6 @@
 /**
  * @internal
  */
-export const ClientSharedValues = {};
+export const ClientSharedValues = {
+  lstatSync: () => {},
+};